// sdks/sdkwork-craw-chat-sdk/sdkwork-craw-chat-sdk-typescript/src/types.ts
export interface CrawChatPreparedMediaAsset {
  mediaAssetId: string;
  createdAsset: MediaAsset;
  completedAsset: MediaAsset;
}

export interface CrawChatMediaUploadSession {
  mediaAssetId: string;
  mediaAsset: MediaAsset;
  bucket: string;
  objectKey: string;
  storageProvider: string;
  uploadMethod: string;
  uploadUrl: string;
  uploadHeaders?: StringMap;
  uploadExpiresInSeconds?: number;
  requestKey?: string;
  deliveryStatus?: 'applied' | 'replayed';
  proofVersion?: string;
}

export interface CrawChatMediaUploadOptions {
  mediaAssetId: string;
  bucket: string;
  objectKey?: string;
  resource: MediaResource;
  body: Exclude<RequestInit['body'], null | undefined>;
  expiresInSeconds?: number;
  checksum?: string;
}

export interface CrawChatUploadedMediaAsset extends CrawChatPreparedMediaAsset {
  asset: MediaAsset;
  url?: string;
  session: CrawChatMediaUploadSession;
  etag?: string;
}